+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 
 * [`PowPersistentSession.Plug.Cookie`] Now supports `:persistent_session_cookie_opts` to customize any options that will be passed on to `Plug.Conn.put_resp_cookie/4`
+* Added mailer rate limitation section to [production checklist guide](guides/production_checklist.md)
 
 ### Bug fixes